Separate promise-related logic into their own files
